🗺️ Atlas: Extract layer registration types
